+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added defaulting for the Cluster description in the `AWSCluster` CR
+
 ## [2.1.0] - 2020-10-29
 
 ### Added