+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added defaulting for the Cluster description in the `AWSCluster` CR
+- Added defaulting for the Cluster DNS domain in the `AWSCluster` CR
 
 ## [2.1.0] - 2020-10-29